Fix all CI lint errors and warnings
- Made AppTheme colors public (primaryLight, accent, errorColor)
- Removed unused _success and _warning fields
- Fixed CalendarScreen passing BuildContext properly
- Removed unused imports across all 15+ files
- Fixed test/widget_test.dart placeholder
- Fixed DoubleExtensions.clampToRange return type
- Removed unused fields in SimulationEngine, AiService, ChatProvider
- Fixed unused variables in career_provider && memory_service
- Removed dart:math, dart:convert, intl unused imports

+++ lib/utils/extensions.dart
@@ -34,7 +34,7 @@ extension IntExtensions on int {
 
 extension DoubleExtensions on double {
   double clampToRange(double min, double max) {
-    return clamp(min, max);
+    return clamp(min, max).toDouble();
   }
 
   String toCurrency({String symbol = '\$', int decimals = 0}) {
